upgraded android native files
